# ordered: fix typo