@lucmq 2024-06-17
### Benchmarks
#### Before
```
goos: linux
goarch: amd64
pkg: rsc.io/ordered
cpu: AMD EPYC 7B13
BenchmarkAppend/Simple-8                34018509               171.1 ns/op            48 B/op          5 allocs/op
PASS
ok      rsc.io/ordered  6.099s
```

##### After
```
goos: linux
goarch: amd64
pkg: rsc.io/ordered
cpu: AMD EPYC 7B13
BenchmarkAppend/Simple-8                78861392                75.96 ns/op            0 B/op          0 allocs/op
PASS
ok      rsc.io/ordered  6.167s
```